requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-yasg==1.21.7

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Validation & Serialization
django-filter==23.5

# Date/Time Handling
python-dateutil==2.8.2

# HTTP Client (for testing)
requests==2.31.0

# Image Processing
Pillow==10.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0 